Add rules for Dublin. Add dont_know answers.
For some rules the user can enter "don't know" and still succeed.
Every answer not n, no, y or yes will be interpreted as "don't know".
Note: For most questions the don't know will be interpreted as a "no" by
the system, for the "don't know" to succeeed the rule need to be very
general and not that country-specifik.